requirements.txt
flask
openai
llama-cpp-python
PyYAML
python-docx
openpyxl
pdfplumber 